requirements.txt
attrs==21.2.0
autopep8==1.6.0
iniconfig==1.1.1
packaging==21.2
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==2.4.7
pytest==6.2.5
pytest-html==3.1.1
pytest-metadata==1.11.0
toml==0.10.2
